fix(export): Corriger export QGIS panic - type i64 pour n_sondages
Le serveur paniquait car n_sondages est INT8 (i64) dans la DB mais était lu comme i32

+++ services/api-geo/src/thematic/export.rs
@@ -133,7 +133,7 @@ pub async fn export_qgis_package(
         let code: String = row.get("code");
         let geom_json: String = row.get("geom_json");
         let value: Option<f64> = row.get("value");
-        let n_sondages: Option<i32> = row.get("n_sondages");
+        let n_sondages: Option<i64> = row.get("n_sondages");
         let adm1: Option<String> = row.get("adm1_name");
         let adm2: Option<String> = row.get("adm2_name");
         let adm3: Option<String> = row.get("adm3_name");
